Imports and Authentication
This project starts with the importation of all require technologies for the structure and functionality of the application. After doing so the authentication and authorization process is started by creating a controllers folder and establishing users by creating a models folder. Bcrypt and JWT is then import to interact with the models created for user.