Finished Routes for Posts and start of front end
The routes for posts including the adding, deleting. and updating have been added along with the corresponding model. Dont similarliy to the users, the posts will include likes, comments, and other criteria that intersects with other data. Dummy data was also inserted for convience but must be deleted and recongfigured so app can be used by client. React and redux was also installed and the front end was started.