Front End File Structure and Redux
I have set the file structure for the front end and doing so has taught me how file names and file interaction is important in organization. I did not realize there is a way to correctly format folders/files for this interaction. I have named folders/files according to their function but theres a better way to do so in order to maintain good coding practices. Also this is my first time working closely with redux and state. I created a state file where all imports are made and following a youtube video I found the importance of having a framework like state in such a project as this. Also redux is a valuable toolkit when doing so.