Navbar Completion and Start of Login/Register
Long time no see, I have finished the Navbar which in all honesty i did not expect all that work to just show that small milstone but it seems that setting foundation for the rest of the page makes the entire application more simple. Now that the navbar is finished I have moved to the login/register page which is using Formik. I'm interested to see how this operates.