Login/Register Finish
I managed to finsih the login and register pages in the Form.jsx. This is definitely something I need to review as I used dropzone and the input of data based from the text fields is very confusing. Also I am getting an error from the navbar from line 43 due to the const fullName which from other comments shows needs to be changed. However when taking it out it causes the app to crash. Other than that the app and database is functioning alright.