POST Widget
I'm still having issues logging/registering in but I'm continuing the project until I do not need the seeded data to see if I can get it to function then. For now I have it locked on the home page where I just added the post widget. Using the same structure as the Form and USerWidget, I added the beginning icons and structure of the posting area for users.