PostWidget and FriendWidget
I finished the MyPostWidget and now Im working on the physically and inidivudal posting between each user id. Similar structure to all the other jsx widgets, I used react components and mui materials to most the adaptive layout. I now need to be careful on how many ids im calling to as likes and comments need to be displayed for each post. I also created the friends widget which can be called on the sidebar which can allow users to follow or unfollow users who posted.